Convert Capabilities to ouroboros.
