+++ Cargo.toml
@@ -27,6 +27,7 @@ nom = { version = "6.0", default-features = false }
 base64 = "0.13"
 chrono = { version = "0.4", default-features = false, features = ["std"]}
 lazy_static = "1.4"
+ouroboros = "0.9.5"
 
 [dev-dependencies]
 lettre = "0.9"
