+++ src/types/capabilities.rs
@@ -1,6 +1,11 @@
-use imap_proto::types::Capability;
+use crate::error::{Error, ParseError};
+use crate::parse::try_handle_unilateral;
+use crate::types::UnsolicitedResponse;
+use imap_proto::{Capability, Response};
+use ouroboros::self_referencing;
 use std::collections::hash_set::Iter;
 use std::collections::HashSet;
+use std::sync::mpsc;
 
 const IMAP4REV1_CAPABILITY: &str = "IMAP4rev1";
 const AUTH_CAPABILITY_PREFIX: &str = "AUTH=";
@@ -30,16 +35,54 @@ const AUTH_CAPABILITY_PREFIX: &str = "AUTH=";
 ///
 /// Client implementations SHOULD NOT require any capability name other than `IMAP4rev1`, and MUST
 /// ignore any unknown capability names.
-pub struct Capabilities(
-    // Note that this field isn't *actually* 'static.
-    // Rather, it is tied to the lifetime of the `ZeroCopy` that contains this `Name`.
-    pub(crate) HashSet<Capability<'static>>,
-);
+#[self_referencing]
+pub struct Capabilities {
+    data: Vec<u8>,
+    #[borrows(data)]
+    #[covariant]
+    pub(crate) capabilities: HashSet<Capability<'this>>,
+}
 
 impl Capabilities {
+    /// Parse the given input into one or more [`Capabilitity`] responses.
+    pub fn parse(
+        owned: Vec<u8>,
+        unsolicited: &mut mpsc::Sender<UnsolicitedResponse>,
+    ) -> Result<Self, Error> {
+        CapabilitiesTryBuilder {
+            data: owned,
+            capabilities_builder: |input| {
+                let mut lines = input;
+                let mut caps = HashSet::new();
+                loop {
+                    match imap_proto::parser::parse_response(lines) {
+                        Ok((rest, Response::Capabilities(c))) => {
+                            lines = rest;
+                            caps.extend(c);
+                        }
+                        Ok((rest, data)) => {
+                            lines = rest;
+                            if let Some(resp) = try_handle_unilateral(data, unsolicited) {
+                                break Err(resp.into());
+                            }
+                        }
+                        _ => {
+                            break Err(Error::Parse(ParseError::Invalid(lines.to_vec())));
+                        }
+                    }
+
+                    if lines.is_empty() {
+                        break Ok(caps);
+                    }
+                }
+            },
+        }
+        .try_build()
+    }
+
     /// Check if the server has the given capability.
     pub fn has<'a>(&self, cap: &Capability<'a>) -> bool {
-        self.0.contains(cap)
+        self.borrow_capabilities().contains(cap)
     }
 
     /// Check if the server has the given capability via str.
@@ -59,16 +102,16 @@ impl Capabilities {
 
     /// Iterate over all the server's capabilities
     pub fn iter(&self) -> Iter<'_, Capability<'_>> {
-        self.0.iter()
+        self.borrow_capabilities().iter()
     }
 
     /// Returns how many capabilities the server has.
     pub fn len(&self) -> usize {
-        self.0.len()
+        self.borrow_capabilities().len()
     }
 
     /// Returns true if the server purports to have no capabilities.
     pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
+        self.borrow_capabilities().is_empty()
     }
 }
